Move ledger config
