+++ packages/apps-config/src/settings/ethereumChains.ts
