+++ packages/apps-config/src/settings/index.ts
@@ -1,6 +1,7 @@
 // Copyright 2017-2020 @polkadot/apps-config authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './ethereumNetworks';
+export * from './ethereumChains';
 export * from './languages';
+export * from './ledgerChains';
 export * from './ss58';
